% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcontext.R
\name{db_connected}
\alias{db_connected}
\title{Check if connected to a database}
\usage{
db_connected(ctx)
}
\arguments{
\item{ctx}{Database context}
}
\description{
Bug: returns TRUE if database connection is expired
}
