% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcontext.R
\name{defdbcontext}
\alias{defdbcontext}
\title{Database context constructor}
\usage{
defdbcontext(host, port, dbname, user, password, auto_disconnect = TRUE)
}
\arguments{
\item{host}{Database host.}

\item{port}{Database port}

\item{dbname}{Database name.}

\item{user}{Database user.}

\item{password}{Database password.}

\item{auto_disconnect}{Should it be closed after usage}
}
\description{
Database context constructor
}
