% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{mdrates}
\alias{mdrates}
\title{Returns the list of zoo objects indexed by features.
Each zoo object contains observations for the symbols requested.}
\usage{
mdrates(
  ctx,
  symbols,
  features = c("open", "high", "low", "close", "volume", "adjusted")
)
}
\description{
Returns the list of zoo objects indexed by features.
Each zoo object contains observations for the symbols requested.
}
\author{
Alexander Dovzhikov
}
