% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\name{security_list}
\alias{security_list}
\title{Returns a list of currently supported securities}
\usage{
security_list(conn)
}
\description{
Returns a list of currently supported securities (as a data frame).
}
\author{
Alexander Dovzhikov
}
