% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsnap.R
\name{mdload}
\alias{mdload}
\title{Market Data Load}
\usage{
mdload(
  ctx,
  symbols,
  asof = NULL,
  features = c("open", "high", "low", "close", "volume", "adjusted")
)
}
\arguments{
\item{ctx}{Database context}

\item{symbols}{A set of symbols to extract, must be specified explicitly.}

\item{asof}{Allows to specify a snapshot to extract, NULL means to latest.}

\item{features}{Features to extract.}
}
\description{
This function loads the specified set of symbols from the database
into an xts object.
}
\author{
Alexander Dovzhikov, \email{alexander.dovzhikov@gmail.com}
}
