% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsnap.R
\name{mdsnap}
\alias{mdsnap}
\title{Market Data Snap}
\usage{
mdsnap(conn)
}
\arguments{
\item{conn}{Database connection or connection pool}
}
\description{
This function snaps current market data snapshot for the symbols universe
and puts them into PostgreSQL database. The universe is configured in the
database itself and needs to be set up manually.
}
